+++ app/controllers/accounts_controller.rb
@@ -1,0 +1,7 @@
+# Helper controller for rodauth acounts
+# Only supports Index
+class AccountsController < ApplicationController
+  def index
+    @accounts = Account.all
+  end
+end