+++ app/helpers/application_helper.rb
@@ -41,7 +41,6 @@ module ApplicationHelper
     end
   end
 
-
   def search_form_tag(url_options, html_options={})
     default_html_options = {
       class: 'search-form',
@@ -54,4 +53,20 @@ module ApplicationHelper
       concat submit_tag("Search")
     end
   end
+
+  def sort_models(models, sort_info)
+    sort_column = sort_info[:column]
+    sort_direction = sort_info[:direction]
+    # models.order("#{sort_column} #{sort_direction}")
+
+    # WORKAROUND: support sorting by non-AR columns
+    # Yes, this is inefficient but our lists are small
+    # Using format(), with leading zeros, to support sorting string, number, or TrueClass
+    models = models.sort_by{|r| format('%010s' % r.send(sort_column).to_s)}
+    if sort_direction.to_s != 'asc'
+      models = models.reverse
+    end
+    models
+  end
+
 end