+++ config/environments/production.rb
@@ -90,4 +90,8 @@ Rails.application.configure do
 
   # Do not dump schema after migrations.
   config.active_record.dump_schema_after_migration = false
+
+  # requires by rodauth
+  fail "Assign default_url_options for production"
+  config.action_mailer.default_url_options = { host: 'registrar.ardenbeachesinc.com'}
 end